fastapi-core-service/requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.18.0
python-dotenv>=0.21.0
langchain>=0.0.300
openai>=0.27.0
sqlalchemy>=1.4
psycopg2-binary>=2.9.0
pymilvus>=2.3.0
python-multipart>=0.0.5
pydantic>=1.9.0